+++ src/sys/unix/io/fd_ext.rs
@@ -38,7 +38,7 @@ pub enum MaybeEventedFd {
 /// # }
 /// ```
 pub trait FileDescExt {
-    /// Registers the underlying primitive OS handle with a `tokio` event loop.
+    /// Attempts to register the underlying primitive OS handle with a `tokio` event loop.
     ///
     /// The resulting type is "futures" aware meaning that it is (a) nonblocking,
     /// (b) will notify the appropriate task when data is ready to be read or written
@@ -49,8 +49,12 @@ pub trait FileDescExt {
     /// `unsafe`ly coping raw file descriptors and registering both copies with
     /// the same `Handle`). Doing so may end up starving one of the copies from
     /// receiving notifications from the event loop.
-    #[deprecated(note = "does not handle regular files, use `into_evented2` instead")]
-    fn into_evented(self, handle: &Handle) -> Result<PollEvented<EventedFileDesc>>;
+    ///
+    /// Note: regular files are not supported by the OS primitives which power tokio
+    /// event loops, and will result in an error on registration. However, since
+    /// regular files can be assumed to always be ready for read/write operations,
+    /// we can handle this usecase by not registering those file descriptors within tokio.
+    fn into_evented(self, handle: &Handle) -> Result<MaybeEventedFd>;
 
     /// Attempts to register the underlying primitive OS handle with a `tokio` event loop.
     ///
@@ -68,9 +72,9 @@ pub trait FileDescExt {
     /// event loops, and will result in an error on registration. However, since
     /// regular files can be assumed to always be ready for read/write operations,
     /// we can handle this usecase by not registering those file descriptors within tokio.
+    #[deprecated(since = "0.2.0", note = "renamed to `into_evented`")]
     fn into_evented2(self, handle: &Handle) -> Result<MaybeEventedFd> where Self: Sized {
-        // FIXME(breaking): remove this default
-        self.into_evented(handle).map(MaybeEventedFd::Registered)
+        self.into_evented(handle)
     }
 
     /// Sets the `O_NONBLOCK` flag on the descriptor to the desired state.
@@ -81,12 +85,7 @@ pub trait FileDescExt {
 }
 
 impl FileDescExt for FileDesc {
-    fn into_evented(self, handle: &Handle) -> Result<PollEvented<EventedFileDesc>> {
-        try!(self.set_nonblock(true));
-        PollEvented::new(EventedFileDesc(self), handle)
-    }
-
-    fn into_evented2(self, handle: &Handle) -> Result<MaybeEventedFd> {
+    fn into_evented(self, handle: &Handle) -> Result<MaybeEventedFd> {
         let ret = if is_regular_file(&self)? {
             MaybeEventedFd::RegularFile(self)
         } else {